requirements.txt
streamlit
openai
faiss-cpu
python-docx
numpy
python-dotenv